Original sewn source code
This is the original discovered source code to
Jared Schiffman's sewn project. This is the
code as it was recovered from

/mas/acg/projects/acApps/sewnApp

and that was in turn taken from

/u/jarfish/projects/sewn/newWorkingSewn

Files are being checked in retroactively
under Jared's name and with a date of
August 1, 1998 which approximates the
original project date.

Files imported by Tom White 6 Apr 2021